fix timer and archived games requet
