+++ components/Board/socketHandler.js
@@ -39,10 +39,10 @@ const handleSocket = (event, setters) => {
       console.log(data);
       break;
     case Actions.MOVE_MADE: {
-      if (!data.isSuccess) {
-        const whiteRemainingTime = data.whiteRemainingTime;
-        const blackRemainingTime = data.blackRemainingTime;
+      const whiteRemainingTime = data.whiteRemainingTime;
+      const blackRemainingTime = data.blackRemainingTime;
 
+      if (!data.isSuccess) {
         console.log(whiteRemainingTime + " " + blackRemainingTime);
         return;
       }
@@ -57,6 +57,8 @@ const handleSocket = (event, setters) => {
       setters.moveList.push(newMove);
       setters.setMoveList([...setters.moveList]);
       setters.setMoveIndex(setters.moveList.length - 1);
+      setters.setWhiteTimer(whiteRemainingTime);
+      setters.setBlackTimer(blackRemainingTime);
       break;
     }
     case Actions.GAME_CREATED:
