+++ components/Profile/ArchivedGamesContainer.jsx
@@ -2,22 +2,99 @@ import { View, Text, TouchableOpacity, StyleSheet } from "react-native";
 import { useState, useEffect } from "react";
 import ArchivedGameCard from "./ArchivedGameCard";
 import SortCriteria from "../../types/SortCriteria";
+import axiosInstance from "../axiosInstance";
 
-const ArchivedGamesContainer = ({ games, playerUsername }) => {
+const ArchivedGamesContainer = ({ playerUsername, timeControl }) => {
   const [sortCriteria, setSortCriteria] = useState(SortCriteria.DESCENDING);
   const [archivedGames, setArchivedGames] = useState([]);
+  // useEffect(() => {
+  //   let sortedGames = [...archivedGames];
+  //   if (sortCriteria === SortCriteria.DESCENDING)
+  //     sortedGames.sort(
+  //       (game, game2) => new Date(game.created) - new Date(game2.created)
+  //     );
+  //   else
+  //     sortedGames.sort(
+  //       (game, game2) => new Date(game2.created) - new Date(game.created)
+  //     );
+  //   setArchivedGames(sortedGames);
+  // }, [sortCriteria, archivedGames]);
+
   useEffect(() => {
-    let sortedGames = [...games];
-    if (sortCriteria === SortCriteria.DESCENDING)
-      sortedGames.sort(
-        (game, game2) => new Date(game.created) - new Date(game2.created)
-      );
-    else
-      sortedGames.sort(
-        (game, game2) => new Date(game2.created) - new Date(game.created)
-      );
-    setArchivedGames(sortedGames);
-  }, [sortCriteria, games]);
+    const token = sessionStorage.getItem("sessionToken");
+    const userId = sessionStorage.getItem("userId");
+
+    const getUserGames = async () => {
+      try {
+        let path = `/archivedGames/${userId}`;
+        if (timeControl) {
+          path = `/archivedGames/${userId}?timeControl=${timeControl}`;
+        }
+
+        const response = await axiosInstance.get(path, {
+          headers: {
+            Authorization: token,
+            userId: userId
+          }
+        });
+        console.log(response.data);
+        //DEBUG:
+        // setArchivedGames(response.data.archivedGames);
+        setArchivedGames([
+          {
+            resultReason: "CHECKMATE",
+            numMoves: 15,
+            timeControl: "BLITZ_5",
+            created: "Oct 12, 2023, 12:13:45 AM",
+            players: [
+              {
+                isWinner: true,
+                playerId: "id1",
+                username: "user1",
+                isWhite: false,
+                rating: 1200
+              },
+              {
+                isWinner: false,
+                playerId: "id2",
+                username: "user2",
+                isWhite: true,
+                rating: 1120
+              }
+            ],
+            id: "670a21a9e135ae2f48a25c00"
+          },
+          {
+            resultReason: "ABORTED",
+            numMoves: 26,
+            timeControl: "BULLET_1",
+            created: "Oct 12, 2024, 12:13:45 AM",
+            players: [
+              {
+                isWinner: false,
+                playerId: "id1",
+                username: "user1",
+                isWhite: true,
+                rating: 1604
+              },
+              {
+                isWinner: true,
+                playerId: "id2",
+                username: "user2",
+                isWhite: false,
+                rating: 120
+              }
+            ],
+            id: "670a21a9e135ae2f48a25c06"
+          }
+        ]);
+      } catch (err) {
+        console.log(err);
+      }
+    };
+
+    getUserGames();
+  }, []);
 
   return (
     <View style={styles.archivedContainer}>
