+++ types/TimeControls.js
@@ -1,9 +1,9 @@
 const TimeControls = Object.freeze({
-    BULLET_1: 60,
-    BULLET_3: 180,
-    BLITZ_5: 300,
-    RAPID_10: 600,
-    CLASSICAL: 1800
+  BULLET_1: 60,
+  BULLET_3: 180,
+  BLITZ_5: 300,
+  RAPID_10: 600,
+  CLASSICAL: 1800
 });
 
 export default TimeControls;
